event-service/requirements.txt
Flask==3.1.2
Flask-GraphQL==2.0.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
SQLAlchemy==2.0.45
PyMySQL==1.1.2
PyJWT==2.10.1
Werkzeug==3.1.4
cryptography
promise
requests
